Remove feature 'bindings'
The main use case was to support custom versions of Godot through a convoluted build process.
With the new feature 'custom-godot', the user can directly generate bindings from a Godot version of choice.
However, godot-rust is not designed to be used without the GDNative API, so there is not much point in keeping that feature.