Upgrade miniserde to 0.1.16 again
Apparently, miniserde versions 0.1.10..=0.1.15 work for rustc 1.59.0, but not for older rustc versions.
Our CI only checks the stable channel with 'cargo +nightly update -Z minimal-versions', so this was not caught.

For the moment, no new CI job is added (to keep minimal dependencies and minimal rustc as orthogonal as possible).
Changing the "minimal dependencies" job to also use MSRV would be an option, but that would make it less easy to see whether a job fails due to rustc or dependency versions.